+++ tests/Feature/Import/FromArrayTest.php
@@ -69,7 +69,7 @@ class FromArrayTest extends TestCase
         $this->assertSame('Description data', $cocktail->description);
         $this->assertSame('Garnish data', $cocktail->garnish);
         $this->assertSame('Laravel', $cocktail->source);
-        $this->assertSame(36.81, $cocktail->abv);
+        $this->assertSame(36.82, $cocktail->abv);
         $this->assertSame($method->id, $cocktail->cocktail_method_id);
         $this->assertSame($glass->id, $cocktail->glass_id);
         $this->assertSame('Tag 1', $cocktail->tags->first()->name);