+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# v3.13.0
+## New
+- Added `volume_ml` property to cocktail schema
+    - Shows approximate total drink volume
+- Added `alcohol_units` property to cocktail schema
+    - Shows approximate drink alcohol units
+- Added `calories` property to cocktail schema
+    - Shows approximate drink calories
+- Added "KindredCocktails" scraper
+
+## Fixes
+- ABV calculation now supports all standard units
+- "CocktailParty" scraper now correctly shows source ingredient string
+- Bar export files now correctly get cleaned when deleting a bar
+- Fixed recipes failing to import when ingredients where missing a category
+
 # v3.12.1
 ## Fixes
 - "CocktailParty" scraper now correctly parses units